* started new Effect file importer system (WIP)  * added new tool funtion for Strings  * implemented Texture API for D3D 11 (not tested yet, WIP)  * added some data mappings for G2Core <-> D3D 11

+++ v2.5/GEAR DX 11/DX11GfxGeneral.cpp
@@ -138,10 +138,10 @@ void SetViewport(G2::rect const& vp)
 	D3D11_VIEWPORT viewport;
 	ZeroMemory(&viewport, sizeof(D3D11_VIEWPORT));
 
-	viewport.TopLeftX = vp.x;
-	viewport.TopLeftY = vp.y;
-	viewport.Width = vp.z;
-	viewport.Height = vp.w;
+	viewport.TopLeftX = (float)vp.x;
+	viewport.TopLeftY = (float)vp.y;
+	viewport.Width = (float)vp.z;
+	viewport.Height = (float)vp.w;
 	//viewport.MinDepth = 0.f;
 	//viewport.MaxDepth = 1.f;
